feat: Initialize core backend microservices for auth, chat, user, recipe, notification, and websocket, updating API Gateway.
