+++ backend/auth-service/prisma/schema.prisma
@@ -1,0 +1,7 @@
+generator client {
+  provider = "prisma-client-js"
+}
+
+datasource db {
+  provider = "postgresql"
+}
