+++ backend/notification-service/prisma/schema.prisma
@@ -1,0 +1,35 @@
+// ==================== NOTIFICATIONS ====================
+generator client {
+  provider = "prisma-client-js"
+}
+
+datasource db {
+  provider = "postgresql"
+}
+
+
+model Notification {
+  id        String           @id @default(cuid())
+  type      NotificationType
+  title     String
+  message   String
+  data      Json?
+  isRead    Boolean          @default(false)
+  createdAt DateTime         @default(now())
+
+  userId String
+//   user   User   @relation(fields: [userId], references: [id], onDelete: Cascade)
+
+  @@index([userId, isRead])
+  @@index([createdAt])
+}
+
+enum NotificationType {
+  NEW_FOLLOWER
+  NEW_FRIEND_REQUEST
+  FRIEND_REQUEST_ACCEPTED
+  NEW_COMMENT
+  NEW_RATING
+  NEW_MESSAGE
+  RECIPE_FAVORITED
+}
