+++ backend/recipe-service/tsconfig.json
@@ -1,0 +1,14 @@
+{
+    "compilerOptions": {
+        "outDir": "dist",
+        "rootDir": "src",
+        "sourceMap": true,
+        "module": "commonjs",
+        "target": "ES2020",
+        "moduleResolution": "node",
+        "esModuleInterop": true
+    },
+    "exclude": [
+        "prisma.config.ts"
+    ]
+}
