+++ backend/user-service/prisma/schema.prisma
@@ -1,0 +1,80 @@
+
+// ==================== UTILISATEURS ====================
+generator client {
+  provider = "prisma-client-js"
+}
+
+datasource db {
+  provider = "postgresql"
+}
+
+
+model User {
+  id         String    @id @default(cuid())
+  email      String    @unique
+  username   String    @unique
+  password   String
+  firstName  String?
+  lastName   String?
+  bio        String?
+  avatarUrl  String?   @default("/default-avatar.png")
+  isOnline   Boolean   @default(false)
+  lastSeenAt DateTime?
+  createdAt  DateTime  @default(now())
+  updatedAt  DateTime  @updatedAt
+
+//   recipes                Recipe[]
+  followers              Follow[]                  @relation("Following")
+  following              Follow[]                  @relation("Followers")
+//   ratings                Rating[]
+//   comments               Comment[]
+//   favorites              Favorite[]
+//   sentMessages           Message[]                 @relation("SentMessages")
+//   conversations          ConversationParticipant[]
+//   notifications          Notification[]
+  friendRequestsSent     FriendRequest[]           @relation("FriendRequestSent")
+  friendRequestsReceived FriendRequest[]           @relation("FriendRequestReceived")
+
+  @@index([email])
+  @@index([username])
+}
+
+// ==================== SYSTÈME D'AMIS ====================
+
+model FriendRequest {
+  id        String              @id @default(cuid())
+  status    FriendRequestStatus @default(PENDING)
+  createdAt DateTime            @default(now())
+  updatedAt DateTime            @updatedAt
+
+  senderId String
+  sender   User   @relation("FriendRequestSent", fields: [senderId], references: [id], onDelete: Cascade)
+
+  receiverId String
+  receiver   User   @relation("FriendRequestReceived", fields: [receiverId], references: [id], onDelete: Cascade)
+
+  @@unique([senderId, receiverId])
+  @@index([senderId])
+  @@index([receiverId])
+}
+
+enum FriendRequestStatus {
+  PENDING
+  ACCEPTED
+  REJECTED
+}
+
+model Follow {
+  id        String   @id @default(cuid())
+  createdAt DateTime @default(now())
+
+  followerId String
+  follower   User   @relation("Followers", fields: [followerId], references: [id], onDelete: Cascade)
+
+  followingId String
+  following   User   @relation("Following", fields: [followingId], references: [id], onDelete: Cascade)
+
+  @@unique([followerId, followingId])
+  @@index([followerId])
+  @@index([followingId])
+}
